Rename index as core. Rename entry as index
