--- src/entry.js
@@ -1,18 +1,0 @@
-'use strict'
-
-const {
-  collectPuts,
-  collectCalls,
-  collectCallsAndPuts,
-  createSagaTestEngine,
-  throwError,
-} = require('./index')
-
-module.exports = {
-  default: createSagaTestEngine,
-  createSagaTestEngine,
-  collectPuts,
-  collectCalls,
-  collectCallsAndPuts,
-  throwError,
-}
